jinja2: support zero-prefixed integers
Jinja2v2 used to support zero-prefixed integers e.g. 01, 02, 03, ...

Jinja2v3 dropped support for this, fair enough, however, our users
haven't been getting warnings for this so haven't had time to adapt.

This PR patches the Jinja2 parser we use in cylc-rose to shave off the
leading zero before we pass the template variable back through to Jinja2
itself. It's not a particularly nice solution, it is intended as a
stopgap to help users to migrate with the intention that we would retire
it ASAP, probably with the move to Jinja2 3.1.